Add consumers for auction events and mapping profiles
- Implement AuctionCreatedConsumer to handle auction creation events, mapping to Item and saving to MongoDB.
- Implement AuctionDeletedConsumer to handle auction deletion events, removing items from MongoDB.
- Implement AuctionUpdatedConsumer to handle auction updates, modifying existing items in MongoDB.
- Create MappingProfiles for AutoMapper to map AuctionCreated and AuctionUpdated to Item.

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -4,6 +4,9 @@ using AuctionService.DTOs;
 using AuctionService.Entities;
 using AutoMapper;
 using AutoMapper.QueryableExtensions;
+using Contracts;
+using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -15,10 +18,12 @@ public class AuctionsController : ControllerBase
 {
     private readonly AuctionDbContext _context;
     private readonly IMapper _mapper;
-    public AuctionsController(AuctionDbContext context, IMapper mapper)
+    private readonly IPublishEndpoint _publishEndpoint;
+    public AuctionsController(AuctionDbContext context, IMapper mapper, IPublishEndpoint publishEndpoint)
     {
         _context = context;
         _mapper = mapper;
+        _publishEndpoint = publishEndpoint;
     }
 
     [HttpGet]
@@ -45,14 +50,18 @@ public class AuctionsController : ControllerBase
         return _mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
     {
         var auction = _mapper.Map<Auction>(createAuctionDto);
-        auction.Seller = "test";
-
+        auction.Seller = User.Identity.Name;
 
         _context.Auctions.Add(auction);
+
+        var newAuction = _mapper.Map<AuctionDto>(auction);
+        await _publishEndpoint.Publish(_mapper.Map<AuctionCreated>(newAuction));
+
         var result = await _context.SaveChangesAsync() > 0;
 
         if (!result)
@@ -60,11 +69,10 @@ public class AuctionsController : ControllerBase
             return BadRequest("Could not save changes to the DB");
         }
 
-        var auctionDto = _mapper.Map<AuctionDto>(auction);
-
-        return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, auctionDto);
+        return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -74,6 +82,11 @@ public class AuctionsController : ControllerBase
             return NotFound();
         }
 
+        if (auction.Seller != User.Identity.Name)
+        {
+            return Forbid();
+        }
+
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
         auction.Item.Year = updateAuctionDto.Year ?? auction.Item.Year;
@@ -81,6 +94,7 @@ public class AuctionsController : ControllerBase
         auction.Item.Mileage = updateAuctionDto.Mileage ?? auction.Item.Mileage;
         auction.Item.ImageUrl = updateAuctionDto.ImageUrl ?? auction.Item.ImageUrl;
 
+        await _publishEndpoint.Publish(_mapper.Map<AuctionUpdated>(auction));
 
         var result = await _context.SaveChangesAsync() > 0;
 
@@ -92,6 +106,7 @@ public class AuctionsController : ControllerBase
         return Ok();
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -100,8 +115,15 @@ public class AuctionsController : ControllerBase
         {
             return NotFound();
         }
+        if (auction.Seller != User.Identity.Name)
+        {
+            return Forbid();
+        }
 
         _context.Auctions.Remove(auction);
+
+        await _publishEndpoint.Publish(new AuctionDeleted { Id = id.ToString() });
+
         var result = await _context.SaveChangesAsync() > 0;
 
         if (!result)
